@@ -1,19 +1,23 @@
 package com.baeldung.httpclient.sec;
 
-import org.apache.http.client.HttpClient;
 import org.apache.http.client.methods.CloseableHttpResponse;
 import org.apache.http.client.methods.HttpGet;
 import org.apache.http.client.protocol.HttpClientContext;
+import org.apache.http.cookie.ClientCookie;
 import org.apache.http.impl.client.BasicCookieStore;
 import org.apache.http.impl.client.CloseableHttpClient;
+import org.apache.http.impl.client.DefaultHttpClient;
 import org.apache.http.impl.client.HttpClientBuilder;
 import org.apache.http.impl.cookie.BasicClientCookie;
 import org.apache.http.protocol.BasicHttpContext;
 import org.apache.http.protocol.HttpContext;
 import com.baeldung.httpclient.ResponseUtil;
+
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 
@@ -25,6 +29,8 @@ public class HttpClientCookieLiveTest {
     private CloseableHttpClient instance;
 
     private CloseableHttpResponse response;
+    
+    private static Logger log = LoggerFactory.getLogger(HttpClientCookieLiveTest.class);
 
     @Before
     public final void before() {
@@ -54,11 +60,15 @@ public class HttpClientCookieLiveTest {
         final BasicCookieStore cookieStore = new BasicCookieStore();
         final BasicClientCookie cookie = new BasicClientCookie("JSESSIONID", "1234");
         cookie.setDomain(".github.com");
+        cookie.setAttribute(ClientCookie.DOMAIN_ATTR, "true");
+
         cookie.setPath("/");
         cookieStore.addCookie(cookie);
-        final HttpClient client = HttpClientBuilder.create().setDefaultCookieStore(cookieStore).build();
-
-        final HttpGet request = new HttpGet("http://www.github.com");
+        
+        DefaultHttpClient client = new DefaultHttpClient();
+        client.setCookieStore(cookieStore);
+        
+        final HttpGet request = new HttpGet("https://www.github.com");
 
         response = (CloseableHttpResponse) client.execute(request);
 
@@ -70,6 +80,7 @@ public class HttpClientCookieLiveTest {
         final BasicCookieStore cookieStore = new BasicCookieStore();
         final BasicClientCookie cookie = new BasicClientCookie("JSESSIONID", "1234");
         cookie.setDomain(".github.com");
+        cookie.setAttribute(ClientCookie.DOMAIN_ATTR, "true");
         cookie.setPath("/");
         cookieStore.addCookie(cookie);
         instance = HttpClientBuilder.create().setDefaultCookieStore(cookieStore).build();
