@@ -1,6 +1,5 @@
 package com.baeldung.servlets;
 
-
 import java.io.IOException;
 
 import javax.servlet.RequestDispatcher;
@@ -18,10 +17,14 @@ public class UpdateServlet extends HttpServlet {
 
         HttpSession session = request.getSession(false);
 
-        session.setAttribute("userName", request.getParameter("userName"));
-        session.setAttribute("age", request.getParameter("age"));
+        if (session != null) {
 
-        request.setAttribute("sessionData", session);
+            session.setAttribute("userName", request.getParameter("userName"));
+            session.setAttribute("age", request.getParameter("age"));
+
+            request.setAttribute("sessionData", session);
+        }
+
         RequestDispatcher requestDispather = request.getRequestDispatcher("update.jsp");
 
         requestDispather.forward(request, response);
